requirements.txt
scikit-surprise
numpy